feat(navbar): implement desktop and mobile navigation components
- Added NavbarDesktopActions for desktop navigation with search, login, cart, and favorites functionality.
- Introduced NavbarLogo for displaying the brand logo with dynamic styling based on the theme.
- Created NavbarMobileActions for mobile navigation, including a search button and cart icon.
- Developed NavbarMobileMenu for mobile menu interactions and navigation links.
- Implemented NavbarNavigation for managing navigation links and active states.
- Added useNavbarConfig hook to manage navbar styles and visibility based on scroll and route conditions.
